Fix grappler crash
