+++ tensorflow/core/grappler/clusters/BUILD
@@ -36,7 +36,7 @@ tf_cuda_library(
     ] + select({
         ":xsmm": ["@libxsmm_archive//:xsmm_avx"],
         "//conditions:default": [],
-    }) + if_dml(["//tensorflow/core/common_runtime/dml:dml_core"]),
+    }) + if_dml(["//tensorflow/core:dml_runtime"]),
 )
 
 tf_cc_test(
